AUTOLOADER FAIL
gave it a try to organize my code , diveded up js, php, and the scripts communicated with db in the db folder etc. Had a autoloader that req all the files. But somehow it didnt workout. Some of the functions didnt connect to the db, Some files didnt seem to have acces to other files etc. Short: it was a mess. So here I revert back to how it was before